Hotfix SimpleInjector NullReference Exception
When using simple injector as DI i got the null reference exception when the SimpleInjector calls its verify method. Reason for the error is this line

 MessageContainer = messageContainerFactory.Create<TMessage>();

in the constructor of ToastNotification class, because the MessageContainerFactory.Create Method depends on HttpContext which is null at the time of verify called mad by DI. The propsed changes fix this issue.